fix: improving the implementation of digital filter to make it easier to maintain in the future
