+++ src/data/data1d/Datum1D.js
@@ -97,9 +97,7 @@ export class Datum1D {
         {
           name: Filters.digitalFilter.id,
           options: {
-            dspfvs: this.info.dspfvs,
-            decim: this.info.decim,
-            grpdly: this.info.grpdly,
+            digitalFilterValue: this.info['digital filter'],
           },
           isDeleteAllow: false,
         },
