+++ src/data/data1d/filter1d/zeroFilling.js
@@ -9,10 +9,10 @@ export function apply(datum1D, size) {
   if (!isApplicable(datum1D)) {
     throw new Error('zeroFilling not applicable on this data');
   }
-
+  let grpdly = datum1D.info['digital filter'];
   let pointsToShift;
-  if (datum1D.info.grpdly > 0) {
-    pointsToShift = Math.floor(datum1D.info.grpdly);
+  if (grpdly > 0) {
+    pointsToShift = Math.floor(grpdly);
   } else {
     pointsToShift = 0;
   }
